requirements.txt
aiogram==2.15
aiohttp==3.7.4.post0
altair==4.1.0
altgraph==0.17
apiai==1.2.3
argon2-cffi==21.1.0
asgiref==3.4.1
astor==0.8.1
astroid==2.8.2
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
auto-py-to-exe==2.10.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
base58==2.1.0
beautifulsoup4==4.10.0
black==21.8b0
bleach==4.1.0
blinker==1.4
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
cfscrape==2.1.1
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
cryptography==35.0.0
cssselect==1.1.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
dicttoxml==1.7.4
distlib==0.3.3
Django==3.2.8
djangorestframework==3.13.1
docopt==0.6.2
docutils==0.17.1
Eel==0.12.4
emoji==1.4.2
entrypoints==0.3
filelock==3.1.0
Fiona @ file:///C:/Users/User/pipwin/Fiona-1.8.20-cp39-cp39-win_amd64.whl
future==0.18.2
GDAL @ file:///C:/Users/User/pipwin/GDAL-3.3.2-cp39-cp39-win_amd64.whl
geocoder==1.38.1
geoip2==4.4.0
gevent==21.8.0
gevent-websocket==0.10.1
gitdb==4.0.7
GitPython==3.1.24
greenlet==1.1.1
h11==0.12.0
icmplib==3.0.1
idna==3.2
importlib-metadata==4.8.1
ip2geotools==0.1.6
IP2Location==8.6.2
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.9.3
jedi==0.18.0
Jinja2==3.0.1
Js2Py==0.71
jsonschema==3.2.0
jupyter-client==7.0.3
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keyring==23.2.1
lazy-object-proxy==1.6.0
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
maxminddb==2.2.0
mccabe==0.6.1
mistune==0.8.4
module-name==0.5.1
multidict==5.1.0
munch==2.5.0
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy @ file:///C:/Users/User/pipwin/numpy-1.21.2%2Bmkl-cp39-cp39-win_amd64.whl
outcome==1.1.0
packaging==21.0
pandas @ file:///C:/Users/User/pipwin/pandas-1.3.3-cp39-cp39-win_amd64.whl
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pefile==2021.5.24
pep8==1.7.1
pickleDB==0.9.2
pickleshare==0.7.5
Pillow==8.3.2
pip-review==1.1.0
pipwin==0.5.1
pkginfo==1.7.1
platformdirs==2.3.0
pluggy==1.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
py==1.10.0
pyarrow==5.0.0
pycodestyle==2.8.0
pycparser==2.20
pydeck==0.7.0
Pygments==2.10.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyjsparser==2.7.1
pylint==2.11.1
Pymem==1.8.5
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyperclip==1.8.2
PyPrind==2.11.3
pyproj @ file:///C:/Users/User/pipwin/pyproj-3.2.1-cp39-cp39-win_amd64.whl
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyquery==1.4.3
pyrsistent==0.18.0
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
pywin32-ctypes==0.2.0
pywinpty==1.1.4
pyzmq==22.3.0
ratelim==0.1.6
readme-renderer==30.0
regex==2021.8.28
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
Rtree @ file:///C:/Users/User/pipwin/Rtree-0.9.7-cp39-cp39-win_amd64.whl
scapy==2.4.5
schedule==1.1.0
selenium==4.0.0
Send2Trash==1.8.0
Shapely @ file:///C:/Users/User/pipwin/Shapely-1.7.1-cp39-cp39-win_amd64.whl
six==1.16.0
smmap==4.0.0
sniffio==1.2.0
socks==0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.2
streamlit==0.88.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tomli==1.2.1
toolz==0.11.1
tornado==6.1
tox==3.24.4
tqdm==4.62.3
traitlets==5.1.0
trio==0.19.0
trio-websocket==0.9.2
twine==3.4.2
typed-ast==1.4.3
typing==3.7.4.3
typing-extensions==3.10.0.2
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.6
validators==0.18.2
virtualenv==20.8.1
watchdog==2.1.5
wcwidth==0.2.5
webencodings==0.5.1
whichcraft==0.6.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wsproto==1.0.0
yarl==1.6.3
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
